Move activist test utils to avoid cyclic dependency
